Updated OpenAI API integration and fixed key issues
