+++ pages/api/chat.js
@@ -1,4 +1,3 @@
-// pages/api/chat.js
 import { createMessage } from '@lib/openai';  // Importing function to interact with OpenAI
 
 export default async function handler(req, res) {
@@ -11,7 +10,7 @@ export default async function handler(req, res) {
         return res.status(400).json({ error: 'Missing required fields: message, thread_id, or username' });
       }
 
-      // Call createMessage function to send the message to the OpenAI API
+      // Call createMessage function to send the message to OpenAI
       const response = await createMessage(thread_id, message);
 
       // Assuming the response has a message field from the assistant
@@ -21,10 +20,11 @@ export default async function handler(req, res) {
       res.status(200).json({ assistantMessage });
 
     } catch (error) {
-      console.error('Error in /api/chat:', error);
-      res.status(500).json({ error: 'Internal Server Error' });
+      console.error('Error in /api/chat:', error.message || error.response?.data);
+      res.status(500).json({ error: 'Internal Server Error: ' + error.message });
     }
   } else {
+    // Method not allowed for other HTTP methods
     res.status(405).json({ error: 'Method Not Allowed' });
   }
 }
